original gpuTile.jl

+++ gpuTile.jl
@@ -17,16 +17,14 @@ gpu_index = length(ARGS) >= 3 ? parse(Int, ARGS[3]) : 0
 CUDA.device!(gpu_index)
 println("Using GPU: ", CUDA.device())
 
-
-
 # Data Setup
-A = rand(Float16, n, n)
-B = rand(Float16, n, n)
+A = rand(Float32, n, n)
+B = rand(Float32, n, n)
 
 dA = CuArray(A)
 dB = CuArray(B)
-dC_tile = CuArray(zeros(Float16, n, n))
-dC_blas = CuArray(zeros(Float16, n, n))
+dC_tile = CuArray(zeros(Float32, n, n))
+dC_blas = CuArray(zeros(Float32, n, n))
 
 
 # Tile-based GPU Kernel
@@ -36,7 +34,7 @@ function gpu_tile_kernel(C, A, B, N, tile_size)
     col = (blockIdx().x - 1) * blockDim().x + threadIdx().x
 
     if row <= N && col <= N
-        acc = Float16(0)
+        acc = 0.0f0
         for kk in 1:tile_size:N
             k_max = min(kk + tile_size - 1, N)
             for k in kk:k_max
@@ -77,8 +75,9 @@ tile_gflops = gflops(n, tile_time)
 
 
 # Run cuBLAS gemm
+
 r_blas = @benchmark begin
-    CUDA.CUBLAS.gemm!('N', 'N', Float16(1.0), dA, dB, Float16(0.0), dC_blas)
+    CUDA.CUBLAS.gemm!('N', 'N', Float32(1.0), dA, dB, Float32(0.0), dC_blas)
     synchronize()
 end samples=5 evals=1
